Move may_segfault
